requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.9
sentence-transformers==2.2.2
faiss-cpu==1.13.2
numpy==1.24.3
scipy==1.11.4
redis==5.0.1
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
